Added some stuff. Starting to like the log writer stuff although the design can still be improved. Hmmm.... if you express feelings in a git commit message, are you a weirdo who talks to himself??

+++ src/Raft.Infrastructure/IO/BufferedSequentialFileWriter.cs
@@ -22,7 +22,7 @@ namespace Raft.Infrastructure.IO
             }
         }
 
-        public void Write(string filePath, int offset, byte[] data)
+        public void Write(string filePath, long offset, byte[] data)
         {
             if (!File.Exists(filePath))
                 throw new FileNotFoundException(string.Format(
